refactor: rename/delete fork pages as we should have only one fork as the AB testing has ended
